h/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.9.0